Fixing bug Undefined offset: 0
When using your package like this for example: 
$disqus->users->listForums()
We get an error of "Undefined offset: 0" that because your $kwargs want the first element of the array $args. But in this case, the args is null and throw and error.
I've added a test to check if the args is null or not before casting.
Hope it's helps.